Parse query string response from reddit auth and send to backend
