+++ app/src/components/Dashboard.jsx
@@ -2,11 +2,8 @@ import React from "react";
 import { withFormik, Form, Field, ErrorMessage } from "formik";
 import * as Yup from "yup";
 import axios from "axios";
-import { useSelector } from "react-redux";
-import { useLocation } from "react-router-dom";
 
-import { DashboardWrapper, ButtonReddit } from "./styled-components";
-import PostCollection from "./PostCollection";
+import { DashboardWrapper } from "./styled-components";
 
 function Dashboard(props) {
   return (
@@ -78,23 +75,4 @@ const DashboardWithFormik = withFormik({
   }
 })(Dashboard);
 
-const DashBoardWithRedditAuth = () => {
-  const redditUrl = useSelector(state => state.redditAuthUrl);
-  const authed = useSelector(state => state.redditAuthed);
-  const location = useLocation();
-
-  console.log(location);
-  return (
-    <>
-      {authed ? (
-        <DashboardWithFormik />
-      ) : (
-        <a href={redditUrl}>
-          <ButtonReddit>Authorize with Reddit</ButtonReddit>
-        </a>
-      )}
-    </>
-  );
-};
-
-export default DashBoardWithRedditAuth;
+export default DashboardWithFormik;
